[IMP] account_move_csv_import: CVS and XLSX/XLS/ODS formats are now fully configurable (colums, start line, sheet)
Add model account.move.import.config to be able to save several configurations. Add a default config.
Move module description to README and improve it.
For ODS parsing, switch to the odsparsator lib, because it is published on pypi (the rows lib is completely outdated on pypi)
For XLSX/XLS/ODS import, add support for date as string, to be more tolerant on bad Excel/ODS files.
Methods that don't need to be public are now private
Merge method clean_strip_pivot() inside method _update_pivot()
FEC import: don't crash when the file is not a text FEC file, but give a proper error message.
Use 'with' instruction to create tempfile
Add header line in demo XLSX/XLS/ODS/CSV files